requirements.txt
python-dotenv==1.1.1
strands-agents==1.1.0
strands-agents-tools==0.2.2
strands-agents-builder==0.1.7